requirements.txt
streamlit==1.41.0
Pillow==11.0.0
requests==2.32.3
streamlit-lottie==0.0.5
